v2.1.1 - Apr 12, 2022 - Fixed float vs int error - SetEnable_Disease_Mortality to 0 by default for vector - Mosquito release update - Updated/standarded sugartrap - a node-level intervention - Accommodating windows users - Standardizedoutdoorrestkill intervention. it cannot be distributed - Updated space spraying - Updated the SimpleBednet intervention to two interventions - Updated inputeir intervention to be added to the campaign automatically - Updated reporters with new parameters for DTkTrunk - Fix Scale_larval_habitats was making node_ids into floats - Added Received_Ivermectin BroadcastEvent to the Ivermectin intervention - Fixed examples to reduce confusion - Updated examples for campaign.set_schema(manifest.schema_path)

+++ emodpy_malaria/interventions/scale_larval_habitats.py
@@ -110,7 +110,7 @@ def add_scale_larval_habitats(campaign, df=None,
                                                        "Factor": float(y)})
 
             if by_node:
-                node_ids = [float(x) for x in gdf['NodeID']]
+                node_ids = [int(x) for x in gdf['NodeID']]
             else:
                 node_ids = None
